Upgraded the fieldbook: - added a navigation bar - added tabs for the routes and the pinmap
