+++ app/app/src/main/res/values/strings.xml
@@ -19,4 +19,7 @@
     <string name="hard">Hard</string>
     <string name="unknown">Unknown</string>
     <string name="completed_text">Completed</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
